requirements.txt
scikit-learn>=1.5.0
pandas>=2.0.0
mecab-python3>=1.0.0
numpy>=1.20.0 
